requirements.txt
streamlit>=1.20.0
tensorflow>=2.12.0
pillow>=9.0.0
numpy>=1.23.0